+++ client/src/fetch/deleteTask.js
@@ -1,11 +1,18 @@
+import catchError from "../utils/catchError";
+import { EmptyValueError, WrongTypeError } from "../utils/customErrors";
+
 const deleteTask = async (id) => {
     try {
+
+        if (!id) throw new EmptyValueError(undefined, { id });
+        if (typeof id !== "number") throw new WrongTypeError("number", id, { id });
+
         const delTask = await fetch(`/tasks/${id}`, {
             method: "DELETE"
         });
         const message = await delTask.json();
     } catch (error) {
-        console.error(error.message);
+        catchError(error)
     }
 }
